move testing_data into test_utils/parkfield and modify imports appropriately
